requirements.txt
# DeepDistill — 多源内容深度蒸馏引擎
# 核心依赖（Docker 镜像使用此文件）

# API 服务
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI
click>=8.0
rich>=13.0

# ASR 语音转文字
faster-whisper>=1.0.0

# OCR 图片文字提取
easyocr>=1.7

# 文档提取
PyPDF2>=3.0
python-docx>=1.0
python-pptx>=0.6
openpyxl>=3.1
beautifulsoup4>=4.12

# AI 分析（兼容 OpenAI 接口）
openai>=1.0
tiktoken>=0.5

# 快速 JSON
orjson>=3.9.0

# 视频分析（OpenCV 核心 + 可选 GPU 模型）
opencv-python-headless>=4.8.0
numpy>=1.24.0
# 可选：pip install scenedetect ultralytics mediapipe
# scenedetect[opencv]>=0.6  # 镜头切割（PySceneDetect）
# ultralytics>=8.0          # 物体检测（YOLOv8）
# mediapipe>=0.10           # 人体姿态（MediaPipe）

# HTTP 客户端（视觉素材生成 API 调用）
httpx>=0.25.0

# Google Docs 导出（Markdown -> HTML -> Google Doc）
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
markdown>=3.5
